feat(linux): Complete Linux Vulkan build - fix all linker errors
Phase 54: Comprehensive Linux x86-64 compatibility fixes for GeneralsXZH.

## D3DX8 Math Compatibility (d3dx8_vulkan_math_compat.h)
- Add D3DXVECTOR3(x,y,z) and D3DXVECTOR4(x,y,z,w) constructors
- Add D3DXMATRIX 16-float constructor for matrix initialization
- Add D3DXVec4Dot() and D3DXVec4Transform() functions
- Re-enable BezierSegment.cpp/h with cross-platform D3DX8 math

## Type System Migrations
- Replace __int64 with int64_t across ~30 files (Win64 compatibility)
- Replace Byte with SignedByte in locomotor/physics code
- Fix signed/unsigned comparison warnings

## STL Compatibility
- Migrate WWMath::Min/Max to std::min/std::max (~40 files)
- Comment out conflicting min/max templates in wwmath.h
- Add <algorithm> includes where needed

## Registry Stubs (phase43_registry_stubs.cpp)
- Implement full RegistryClass interface matching registry.h
- Add Exists(), Get_Int(), Set_Int(), Get_String(), Set_String()
- Add Get_Binary(), Set_Binary(), Delete_Value(), Delete_Key()
- Return sensible defaults for non-Windows platforms

## WW3D Wrapper (vulkan_ww3d_wrapper.cpp)
- Fix WW3D::Render() signature: SceneClass*, CameraClass*
- Fix WW3D::Flush() signature: RenderInfoClass&
- Add TextureFilterClass stubs: Apply(), Set_Default_Min/Mag_Filter()
- Add missing WW3D namespace functions

## glibc 2.40+ Compatibility (stringex.h)
- Add conditional declarations for wcslcpy/wcslcat
- Prevent conflicts with glibc's new wchar.h definitions
- Use __GLIBC_PREREQ macro for version detection

## GameSpy SDK Fixes
- Remove duplicate globals from phase43_5_gamespy_globals.cpp
- Variables moved: TheGameSpyGame, TheGameSpyInfo, GameSpyColor,
  parseWebpageURLDefinition, and 10+ other duplicates
- Add cmake/gamespy.cmake patches for min/max conflicts

## Build System Updates
- Fix OpenAL linking order in CMakeLists.txt
- Update library dependencies for Linux builds
- Add proper include paths for cross-platform headers

## Header Reorganization
- Rename debug.h to debug_renamed.h (conflict resolution)
- Add CrossPlatformOSDisplay.cpp for display abstraction
- Update include guards and paths

Build verified: 179MB ELF x86-64 executable generated successfully.

Refs: Phase 54 - Linux Vulkan Port

+++ Core/Libraries/Source/WWVegas/WW3D2/segline.cpp
@@ -270,7 +270,7 @@ void SegmentedLineClass::Set_Merge_Abort_Factor(float factor)
 
 void SegmentedLineClass::Set_Subdivision_Levels(unsigned int levels)
 {
-	MaxSubdivisionLevels = std::min(levels, MAX_SEGLINE_SUBDIV_LEVELS);
+	MaxSubdivisionLevels = std::min(static_cast<int>(levels), MAX_SEGLINE_SUBDIV_LEVELS);
 
 	Invalidate_Cached_Bounding_Volumes();
 }
